Disable like/delete while in progress.
The like and delete icons are now disabled
 while the fetch to the backend is in pro-
 gress, to prevent API call spamming to
 some degree.

+++ Webservices-consumer/src/components/MovieIconCard.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { computed } from "vue";
+import { ref, type Ref } from "vue";
 import type { Movie } from "@/code/interfaces"
 
 import MovieCard from './MovieCard.vue';
@@ -18,6 +19,18 @@ const emit = defineEmits<{
     (e: "delete", movie: Movie): void;
 }>();
 
+/** Whether a 'like' call is in progress */
+const is_liking: Ref<boolean> = ref(false);
+const class_like = computed(() => {
+  return is_liking.value ? "" : "icon-like"
+});
+
+/** Whether a 'delete' call is in progress */
+const is_deleting: Ref<boolean> = ref(false);
+  const class_delete = computed(() => {
+  return is_deleting.value ? "" : "icon-delete"
+});
+
 /** The color of the like icon, based on movie like status */
 const like_icon_fill = computed(() => {
     return (props.movie.liked !== undefined && props.movie.liked === true) ? "fill: green;" : "";
@@ -26,29 +39,47 @@ const like_icon_fill = computed(() => {
 /** The like/unlike functionality event trigger */
 async function onClickLike() {
   if (props.movie.liked === undefined) return;
+  if (is_liking.value) return;
 
+  is_liking.value = true;
   const method: string = props.movie.liked ? "DELETE" : "PUT";
   fetch(`${import.meta.env.VITE_APP_API_BASE_URL}/likes/${props.movie.id}`, {
     method: method,
     credentials: "same-origin"
   })
   .then(async (response) => {
-    if (response.status >= 400) return;
+    if (response.status >= 400) {
+      is_liking.value = false;
+      return;
+    }
 
     emit("like", props.movie);
+    is_liking.value = false;
+  }).catch(() => {
+    is_liking.value = false;
   });
 }
 
 /** The delete functionality event trigger */
 async function onClickDelete() {
+  if (is_deleting.value) return;
+
+  is_deleting.value = true;
   fetch(`${import.meta.env.VITE_APP_API_BASE_URL}/movies/${props.movie.id}`, {
     method: "DELETE",
     credentials: "same-origin"
   })
   .then(async (response) => {
-    if (response.status >= 400) return;
+    if (response.status >= 400) {
+      is_deleting.value = false;
+      return;
+    }
 
     emit("delete", props.movie);
+    is_deleting.value = false;
+  })
+  .catch(() => {
+    is_deleting.value = false;
   });
 }
 </script>
@@ -58,11 +89,11 @@ async function onClickDelete() {
     <template v-slot:middle>
         <!-- Icons to provide extra functionality -->
         <div class="movie-card-icons">
-        <ThumbsUpLogo class="icon-like"
+        <ThumbsUpLogo :class="class_like"
           :style="like_icon_fill"
           @click="onClickLike"
         />
-        <TrashcanLogo class="icon-delete"
+        <TrashcanLogo style="margin-left: 1rem;" :class="class_delete"
           @click="onClickDelete"
         />
         </div>
@@ -71,10 +102,6 @@ async function onClickDelete() {
 </template>
 
 <style scoped>
-.icon-delete {
-    margin-left: 1rem;
-}
-
 .movie-card-icons {
   margin-left: 1rem;
   margin-right: 1rem;
@@ -86,7 +113,7 @@ async function onClickDelete() {
         cursor: pointer;
     }
 
-    .icon-like {
+    .icon-like:hover {
         cursor: pointer;
     }
 }
